pyproject.toml
[build-system]
requires = [
    "setuptools >= 68.1.2",
    "versioningit >= 2.2.1",
]
build-backend = 'setuptools.build_meta'

[project]
name = "broadbean"
maintainers = [{name = "QCoDeS Core Developers", email = "qcodes-support@microsoft.com"}]
description = "Package for easily generating and manipulating signal pulses."
keywords = ["Pulsebuilding signal processing arbitrary waveforms"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]
license = {text = "MIT"}
urls = {Homepage = "https://github.com/QCoDeS/broadbean"}
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.12.1",
    "matplotlib",
    "schema",
    "versioningit>=2.2.1",
]
dynamic = ["version"]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.optional-dependencies]
test = [
    "pytest>=9.0.0",
    "pytest-cov>=3.0.0",
    "coverage[toml]>=6.2",
    "mypy>=0.960",
    "types-pytz>=2021.3.0",
    "hypothesis>=5.49.0",
]
docs = [
    "sphinx",
    "furo",
    "nbsphinx",
    "sphinx-jsonschema",
    "ipykernel",
]
designer = [
    "django>=4.2",
    "pyyaml",
]
hardware = [
    "qcodes",
    "pyvisa",
]

[project.scripts]
broadbean-designer = "broadbean.designer:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"broadbean.designer" = [
    "templates/**/*",
    "static/**/*",
    "migrations/*.py",
    "*.csv",
]


[tool.pytest]
minversion = "9.0"
addopts = ["-ra", "-q"]

[tool.mypy]
show_error_codes = true
enable_error_code = "ignore-without-code"
warn_unused_ignores = true
warn_unused_configs = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "schema"
]
ignore_missing_imports = true

[tool.ruff]

extend-include = ["*.ipynb"]

[tool.ruff.lint]
extend-select = ["UP", "ISC", "I"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["broadbean"]

[tool.setuptools.cmdclass]
sdist = "versioningit.cmdclass.sdist"
build_py = "versioningit.cmdclass.build_py"

[tool.versioningit]
default-version = "0.16.0"

[tool.versioningit.format]
distance = "{next_version}.dev{distance}+{branch}.{vcs}{rev}"
dirty = "{next_version}.dev{distance}+{branch}.{vcs}{rev}.dirty"
distance-dirty = "{next_version}.dev{distance}+{branch}.{vcs}{rev}.dirty"

[tool.versioningit.vcs]
method = "git"
match = ["v*"]

[tool.versioningit.onbuild]
source-file = "src/broadbean/_version.py"
build-file = "broadbean/_version.py"

[tool.darker]
isort = false
